Connect Proxy to Gui
The proxy.py script sends intercepted requests to main.py, which displays them in the GUI.
The GUI then sends back the user's decision (forward)  through the socket,

When the user presses the "Forward" button in the GUI, the request is immediately forwarded to its destination. This action is triggered by the socket message that indicates the user's decision to proceed with the request.

The GUI dynamically updates to show the intercepted requests, including their method, URL, and content, allowing users to inspect the details before making a decision.